maj3

+++ controllers/post.js
@@ -41,7 +41,6 @@ exports.getAllPost =(req, res, next) => {
 
 exports.getOnePost =(req, res, next) => {
   const idPost = req.params.idPost;
-  console.log(idPost)
   Post.findOne({ _id: idPost })
   .then((post) => {
     if (post) {
@@ -85,30 +84,34 @@ exports.ModifyOnePost = async (req, res, next) => {
     `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     :"");  
   const user = await User.findOne({_id: req.auth.userId})
-    console.log(user)
+ 
   Post.findOne({ _id: idPost })
-      .then(post => {
-        if(newImageContentUrl !== ""){
-          const filenameDelete = post.imageContentUrl.split("/images/")[1];
-          fs.unlink(`images/${filenameDelete}`, () => {
-            Post.updateOne( {_id: idPost}, {$set:{content: newContent, imageContentUrl: newImageContentUrl}} )
+       .then(post => {
+        if(JSON.stringify(post.userId) === JSON.stringify(user._id) || user.isAdmin){
+          if(newImageContentUrl !== ""){
+            const filenameDelete = post.imageContentUrl.split("/images/")[1];
+            fs.unlink(`images/${filenameDelete}`, () => {
+              Post.updateOne( {_id: idPost}, {$set:{content: newContent, imageContentUrl: newImageContentUrl}} )
+                .then(() => {
+                  Post.findOne({_id: idPost }).then((postModified)=> res.status(200).json(postModified))
+                })
+                .catch((error) => {console.log(error);
+                  res.status(400).json({ message: 'Impossible de modifier '})}
+                );
+            })
+          }else{
+            Post.updateOne( {_id: idPost}, {$set:{content: newContent}} )
               .then(() => {
                 Post.findOne({_id: idPost }).then((postModified)=> res.status(200).json(postModified))
               })
               .catch((error) => {console.log(error);
                 res.status(400).json({ message: 'Impossible de modifier '})}
               );
-          })
-        }else{
-          Post.updateOne( {_id: idPost}, {$set:{content: newContent}} )
-            .then(() => {
-              Post.findOne({_id: idPost }).then((postModified)=> res.status(200).json(postModified))
-            })
-            .catch((error) => {console.log(error);
-              res.status(400).json({ message: 'Impossible de modifier '})}
-            );
+          }
+        } else{
+          res.status(401).json({error: 'erreur auth'})
         }
-      })
+      }) 
       .catch((error) =>{console.log(error);
         res.status(500).json({ error: "Une erreur s'est produite !" })}
       );
@@ -116,6 +119,7 @@ exports.ModifyOnePost = async (req, res, next) => {
 
 exports.deleteOnePost =(req, res, next) =>{
   const idPost = req.params.idPost;
+  
   Post.findOne({ _id: idPost })
   .then(post => {
     if (post.imageContentUrl) {
